requirements.txt
aiogram~=2.15
PyYAML~=5.4.1
SQLAlchemy~=1.4.25